:label: seperated custom hook from modal to enhance readablility
